change file name and make models in diff diff files
