+++ controllers/auth/index.js
@@ -1,4 +1,4 @@
-const { UserModel } = require("../../db/Schema");
+const { UserModel } = require("../../db/models/user");
 const bcrypt = require("bcryptjs");
 const _ = require('lodash');
 
@@ -9,18 +9,16 @@ const verifyAuthToken = async (req, res) => {
 
 const SignUp = async (req, res) => {
   try {
+    
     let toAddUser = req.body;
-
     const hasUser = await UserModel.findOne({
       email: toAddUser.email.trim(),
     });
-
     if (hasUser) {
       return res
         .status(404)
         .send({ message: "User with this email already exist" });
     }
-
     const User = {
       ...toAddUser,
     };
@@ -30,16 +28,13 @@ const SignUp = async (req, res) => {
       if (e) {
         return res.status(404).send({ message: e });
       }
-
       const token = await newUser.generateAuthToken();
-  
       return res
         .status(200)
         .send({
           message: "User registered successfully",
           data: { ..._.pick(toAddUser, [ '_id', 'name', 'email' ]), token: token },
         });
-
     });
 
   } catch (e) {
@@ -49,26 +44,26 @@ const SignUp = async (req, res) => {
 
 const SignIn = async (req, res) => {
   try {
+   
     const { email, password } = req.body;
-    
+   
     const User = await UserModel.findOne({ email });
-
     if (!User) {
       return res.status(404).send({ message: "User doesn't exist" });
     }
-
+   
     const isPasswordSame = await bcrypt.compare(password, User.password);
-
     if (!isPasswordSame) {
       return res.status(400).send({ message: "invalid password" });
     }
-
+   
     const token = await User.generateAuthToken();
-
+   
     res.status(200).send({
       message: "User logged in successfully",
       data: { ..._.pick(User, [ '_id', 'name', 'email' ]), token: token },
     });
+    
   } catch (e) {
     res.status(404).send({ message: e.message });
   }
