+++ env.js
@@ -2,3 +2,4 @@ require("dotenv").config();
 
 exports.SECRET_KEY = process.env.SECRET_KEY;
 exports.DATABASE = process.env.DATABASE;
+exports.PORT = process.env.PORT;
