+++ router/auth.js
@@ -1,16 +1,19 @@
 const express = require("express");
 const router = express.Router();
 
-const { Authenticate } = require("../middleware/Authenticate")
-const { Validator } = require("../middleware/Validator")
-
-const { SignIn, SignUp, verifyAuthToken } = require("../controllers/auth/index");
+const { Authenticate } = require("../middleware/authenticate")
+const { Validator } = require("../middleware/validator")
 const { SignUpValidation ,LoginValidation} = require("../controllers/auth/validation");
 
+const { 
+    SignIn, 
+    SignUp, 
+    verifyAuthToken 
+} = require("../controllers/auth/index");
 
 
-router.get("/verifyToken", Authenticate, verifyAuthToken);
 router.post("/signup",SignUpValidation,Validator, SignUp);
 router.post("/login",LoginValidation,Validator, SignIn);
+router.get("/verifyToken", Authenticate, verifyAuthToken);
 
 module.exports = router;
